requirements.txt
holidays
scikit-learn
pandas
matplotlib
numpy
aiohttp
astral
fastapi[standard]
lightgbm
entsoe-py

# Testing dependencies
pytest
pytest-asyncio
httpx